delete post from index bug fixed

+++ admin/includes/view_all_users.php
@@ -115,16 +115,19 @@ if(isset($_GET['change_to_admin'])) {
 
 if(isset($_GET['delete'])) {
     
-    $the_user_id = $_GET['delete'];
+    if($_SESSION['role'] == 'admin'){
+
+            $the_user_id =  mysqli_real_escape_string($connection, $_GET['delete']);
     
     
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+            $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+            
+            $delete_query = mysqli_query($connection, $query);
+            header("Location: users.php");
+        }
     
-    $delete_query = mysqli_query($connection, $query);
     
-    if(!headers_sent()){
-        header("Location: users.php");
-    }
+   
 }                                
                                 
 ?>
